# openapi.yaml
openapi: 3.1.0
info:
  title: Рядом
  version: 0.1.0
paths:
  /api/v1/auth/max:
    post:
      tags:
      - identity
      summary: Login With Max
      operationId: login_with_max_api_v1_auth_max_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaxLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/dev-login:
    post:
      tags:
      - identity
      summary: Dev Login
      operationId: dev_login_api_v1_auth_dev_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/review-login:
    post:
      tags:
      - identity
      summary: Review Login
      operationId: review_login_api_v1_auth_review_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/dev/reset:
    post:
      tags:
      - identity
      summary: Dev Reset
      operationId: dev_reset_api_v1_dev_reset_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: integer
                type: object
                title: Response Dev Reset Api V1 Dev Reset Post
      security:
      - HTTPBearer: []
  /api/v1/me:
    get:
      tags:
      - identity
      summary: Me
      operationId: me_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/max_assist__modules__identity__schemas__MeOut'
      security:
      - HTTPBearer: []
  /api/v1/me/recording-consent:
    post:
      tags:
      - identity
      summary: Agree To Recording
      operationId: agree_to_recording_api_v1_me_recording_consent_post
      responses:
        '204':
          description: Successful Response
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    DevLoginRequest:
      properties:
        user_key:
          type: string
          title: User Key
      type: object
      required:
      - user_key
      title: DevLoginRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MaxLoginRequest:
      properties:
        init_data:
          type: string
          title: Init Data
      type: object
      required:
      - init_data
      title: MaxLoginRequest
    MeCounters:
      properties:
        drafts:
          type: integer
          title: Drafts
        active_assist_sessions:
          type: integer
          title: Active Assist Sessions
        trusted_helpers:
          type: integer
          title: Trusted Helpers
        helping_for:
          type: integer
          title: Helping For
      type: object
      required:
      - drafts
      - active_assist_sessions
      - trusted_helpers
      - helping_for
      title: MeCounters
...
    ReviewLoginRequest:
      properties:
        login:
          type: string
          title: Login
        password:
          type: string
          title: Password
      type: object
      required:
      - login
      - password
      title: ReviewLoginRequest
...
    StaffOut:
      properties:
        role:
          type: string
          title: Role
        organization:
          type: string
          title: Organization
        position:
          anyOf:
          - type: string
          - type: 'null'
          title: Position
        verified:
          type: boolean
          title: Verified
      type: object
      required:
      - role
      - organization
      - position
      - verified
      title: StaffOut
...
    TokenOut:
      properties:
        access_token:
          type: string
          title: Access Token
        expires_at:
          type: string
          format: date-time
          title: Expires At
        user:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - access_token
      - expires_at
      - user
      title: TokenOut
...
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        display_name:
          type: string
          title: Display Name
        photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Url
        staff:
          anyOf:
          - $ref: '#/components/schemas/StaffOut'
          - type: 'null'
      type: object
      required:
      - id
      - display_name
      - photo_url
      - staff
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    max_assist__modules__identity__schemas__MeOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        first_name:
          type: string
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        display_name:
          type: string
          title: Display Name
        photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Url
        staff:
          anyOf:
          - $ref: '#/components/schemas/StaffOut'
          - type: 'null'
        recording_consent:
          type: boolean
          title: Recording Consent
        counters:
          $ref: '#/components/schemas/MeCounters'
      type: object
      required:
      - id
      - first_name
      - last_name
      - display_name
      - photo_url
      - staff
      - recording_consent
      - counters
      title: MeOut
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer